refactor(utils): co-locate isSqliteBusyError in pi-utils/sqlite
Move the busy-family classifier from @oh-my-pi/pi-ai/auth-storage next
to isSqliteCorruptError, so both SQLite result-code classifiers live in
one module. Every caller migrates to the new import; the pi-ai barrel
no longer exports the symbol (breaking for external consumers, noted in
the changelog). Classifier unit tests move to packages/utils/test.

Stacked on the corrupt-store-latch branch, which created
packages/utils/src/sqlite.ts.
