+++ packages/utils/test/sqlite.test.ts
@@ -1,0 +1,32 @@
+import { describe, expect, test } from "bun:test";
+import { isSqliteBusyError } from "../src/sqlite";
+
+interface SqliteBusyShape extends Error {
+	code: string;
+	errno: number;
+}
+
+function makeBusyError(code: string, errno: number): SqliteBusyShape {
+	const err = new Error("database is locked") as SqliteBusyShape;
+	err.code = code;
+	err.errno = errno;
+	return err;
+}
+
+describe("isSqliteBusyError", () => {
+	test("recognizes every documented BUSY family code", () => {
+		expect(isSqliteBusyError(makeBusyError("SQLITE_BUSY", 5))).toBe(true);
+		expect(isSqliteBusyError(makeBusyError("SQLITE_BUSY_RECOVERY", 261))).toBe(true);
+		expect(isSqliteBusyError(makeBusyError("SQLITE_BUSY_SNAPSHOT", 517))).toBe(true);
+		expect(isSqliteBusyError(makeBusyError("SQLITE_BUSY_TIMEOUT", 773))).toBe(true);
+	});
+
+	test("rejects non-BUSY codes and non-error values", () => {
+		expect(isSqliteBusyError(makeBusyError("SQLITE_LOCKED", 6))).toBe(false);
+		expect(isSqliteBusyError(makeBusyError("SQLITE_CORRUPT", 11))).toBe(false);
+		expect(isSqliteBusyError(new Error("plain"))).toBe(false);
+		expect(isSqliteBusyError(null)).toBe(false);
+		expect(isSqliteBusyError(undefined)).toBe(false);
+		expect(isSqliteBusyError("SQLITE_BUSY")).toBe(false);
+	});
+});
